Refactor getExportPhrase to get num descendants from exported content rather than a full traversal.
