+++ src/util/index.ts
@@ -27,13 +27,13 @@ export { equalThoughtSorted } from './equalThoughtSorted'
 export { equalThoughtValue } from './equalThoughtValue'
 export { escapeRegExp } from './escapeRegExp'
 export { escapeSelector } from './escapeSelector'
+export { exportPhrase } from './exportPhrase'
 export { formatNumber } from './formatNumber'
 export { getAlgoliaApiKey } from './getAlgoliaApiKey'
 export { getCaretPositionDetails } from './getCaretPositionDetails'
 export { getContextMap } from './getContextMap'
 export { getDepth } from './getDepth'
 export { getElementPaddings } from './getElementPaddings'
-export { getExportPhrase } from './getExportPhrase'
 export { getPublishUrl } from './getPublishUrl'
 export { getTextContentFromHTML } from './getTextContentFromHTML'
 export { getUserRef } from './getUserRef'
